requirements.txt
pandas
numpy
scikit-learn
joblib
flask
flask-cors
bcrypt
PyJWT